+++ example/src/Examples/Aurora/components/Cubic.tsx
@@ -2,6 +2,8 @@ import React from "react";
 import type { AnimationValue, Vector } from "@shopify/react-native-skia";
 import { Line, Paint, Circle } from "@shopify/react-native-skia";
 
+import { bilinearInterpolate } from "./Math";
+
 interface CubicProps {
   vertices: AnimationValue<Vector[]>;
   index: number;
@@ -9,9 +11,20 @@ interface CubicProps {
   c2: AnimationValue<Vector>;
   c3: AnimationValue<Vector>;
   c4: AnimationValue<Vector>;
+  colors: number[];
+  size: Vector;
 }
 
-export const Cubic = ({ vertices, index, c1, c2, c3, c4 }: CubicProps) => {
+export const Cubic = ({
+  vertices,
+  index,
+  c1,
+  c2,
+  c3,
+  c4,
+  colors,
+  size,
+}: CubicProps) => {
   return (
     <>
       <Line
@@ -26,7 +39,11 @@ export const Cubic = ({ vertices, index, c1, c2, c3, c4 }: CubicProps) => {
         p1={() => c2.value}
         p2={() => c4.value}
       />
-      <Circle c={() => vertices.value[index]} r={16} color="black">
+      <Circle
+        c={() => vertices.value[index]}
+        r={16}
+        color={() => bilinearInterpolate(colors, size, vertices.value[index])}
+      >
         <Paint style="stroke" strokeWidth={4} color="white" />
       </Circle>
       <Circle c={() => c1.value} r={10} color="white" />
